An attempt at fixing alignment of disciplinary entries.
In cases where the number of discipline cases on the "Discipline History" tab matches with exactly one case once clicked into the allegations/charges number, this matches those entries up.

However, this doesn't work for cases where the "Discipline History" tab and the data after clicking on the number have a different number of cases, which happens for Felix A Paulino (tax ID: 962002).